refactor auth stories
